+++ Src/Cocoa/CMCocoaTimer.m
@@ -107,7 +107,7 @@ UInt32 archGetHiresTimer()
 	UInt32 ticks;
 	struct timeval now;
 	gettimeofday(&now, NULL);
-	ticks = (now.tv_sec - start.tv_sec) * 1000 + (now.tv_usec - start.tv_usec) / 1000;
+	ticks = (UInt32)((now.tv_sec - start.tv_sec) * 1000 + (now.tv_usec - start.tv_usec) / 1000);
     
     return ticks;
 }
